# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/menus": {
            "get": {
                "description": "获取菜单列表（支持分页、搜索、排序）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取菜单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（搜索菜单名称）",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式：asc/desc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "菜单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.CreateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_MenuResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
        "/menus/:id": {
            "get": {
                "description": "根据ID获取菜单详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取菜单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_MenuResponse"
                        }
                    },
                    "404": {
                        "description": "菜单不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "put": {
                "description": "更新菜单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.UpdateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_MenuResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "404": {
                        "description": "菜单不存在",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_result.ResponseResult-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_lyj404_gin-api-template_domain_dto.CreateMenuRequest": {
            "type": "object",
            "required": [
                "name",
                "resource_id"
            ],
            "properties": {
                "component": {
                    "description": "前端组件路径",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "is_visible": {
                    "description": "是否显示",
                    "type": "boolean"
                },
                "name": {
                    "description": "菜单名称（必填）",
                    "type": "string"
                },
                "order_num": {
                    "description": "排序序号",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "path": {
                    "description": "前端路由路径",
                    "type": "string"
                },
                "resource_id": {
                    "description": "关联资源ID（必填）",
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.MenuResponse": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.MenuResponse"
                    }
                },
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "order_num": {
                    "type": "integer"
                },
                "parent_id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "total_page": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_dto.UpdateMenuRequest": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "前端组件路径",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "is_visible": {
                    "description": "是否显示",
                    "type": "boolean"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "order_num": {
                    "description": "排序序号",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "path": {
                    "description": "前端路由路径",
                    "type": "string"
                },
                "resource_id": {
                    "description": "关联资源ID",
                    "type": "integer"
                },
                "status": {
                    "description": "状态：enabled/disabled",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_MenuResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.MenuResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-github_com_lyj404_gin-api-template_domain_dto_PaginationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_lyj404_gin-api-template_domain_dto.PaginationResponse"
                        }
                    ]
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
        "github_com_lyj404_gin-api-template_domain_result.ResponseResult-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "@Description 响应状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "@Description 响应数据",
                    "type": "string"
                },
                "message": {
                    "description": "@Description 响应信息",
                    "type": "string"
                }
            }
        },
...
    }
}